backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
gitpython==3.1.40

# ML/AI Dependencies
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3

# Optional: CUDA support
# torch==2.1.1+cu118 -f https://download.pytorch.org/whl/torch_stable.html

# Database (for production)
sqlalchemy==2.0.23
alembic==1.13.0

# Caching
redis==5.0.1

# API clients
anthropic==0.7.7
openai==1.3.7

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
